Remove absolute path to true
Some systems do not have 'true' in /bin/true; trust PATH to find it.

Some BSD systems, MacOS especially, have it in /usr/bin/true. It makes
sense to use PATH to find it. I've no reason to distrust the shell
built-in true, and we could just use that, but I unset it anyway (with
enable) to use the system binary, just in case.